+++ apps/user/src/services/form/api.ts
@@ -1,0 +1,6 @@
+import { maru } from '@/apis/instance/instance';
+
+const submitDraftForm = async () => {
+    const { data } = await maru.post('/form');
+    return { data };
+};
